+++ src/main/java/racingcar/racinggame/car/Cars.java
@@ -1,4 +1,4 @@
-package racingcar.io.domain;
+package racingcar.racinggame.car;
 
 import java.util.HashSet;
 import java.util.List;
