+++ src/main/java/racingcar/util/RandomNumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcar.io.domain;
+package racingcar.util;
 
 import java.util.ArrayList;
 import java.util.List;
